sstables: mc: sign-extend delta local_deletion_time and delta ttl
Follow Cassandra's encoding so that values that are less than the
baseline encoding_stats will wrap-around in 64-bits rather tham 32.

Fixes #4074
Refs #3353

Message-Id: <20190109192703.18371-1-bhalevy@scylladb.com>

+++ sstables/liveness_info.hh
@@ -52,10 +52,10 @@ public:
     void set_timestamp(uint64_t delta) {
         _timestamp = parse_timestamp(_header, delta);
     }
-    void set_ttl(uint32_t delta) {
+    void set_ttl(uint64_t delta) {
         _ttl = parse_ttl(_header, delta);
     }
-    void set_local_deletion_time(uint32_t delta) {
+    void set_local_deletion_time(uint64_t delta) {
         _local_deletion_time = parse_expiry(_header, delta);
     }
     api::timestamp_type timestamp() const { return _timestamp; }
